Moved tests for test scenario config to separate test file
